ln-simln-jamming: move reputation_interceptor into dedicated file
